Showing marketData in WatchListSymbolDetailsFragment.kt

+++ app/src/main/java/com/finsol/tech/data/model/Market.kt
@@ -4,7 +4,7 @@ import android.os.Parcelable
 import kotlinx.parcelize.Parcelize
 
 @Parcelize
-data class Market(val securityID: String, val askPrice: ArrayList<ArrayList<Int>>, val bidPrice: ArrayList<ArrayList<Int>>) :
+data class Market(val securityID: String, val askPrice: ArrayList<ArrayList<Float>>, val bidPrice: ArrayList<ArrayList<Float>>) :
     Parcelable
