+++ app/src/main/java/com/finsol/tech/presentation/watchlist/WatchListSymbolDetailsFragment.kt
@@ -112,12 +112,12 @@ class WatchListSymbolDetailsFragment: BaseFragment() {
     private fun updateBidOfferViewsData(marketDetails: Market) {
         marketDetails.askPrice.forEachIndexed { index, element ->
             offerViews[index].view1.text = element[0].toString()
-            offerViews[index].view3.text = element[1].toString()
+            offerViews[index].view3.text = element[1].toInt().toString()
         }
 
         marketDetails.bidPrice.forEachIndexed { index, element ->
             bidViews[index].view1.text = element[0].toString()
-            bidViews[index].view3.text = element[1].toString()
+            bidViews[index].view3.text = element[1].toInt().toString()
         }
     }
